--- src/Tools/RapidResponseDesk/Advanced/AdvancedSection.scss
@@ -1,1 +1,0 @@
-@import '../../../variables.module.scss';
