--- src/Tools/RapidResponseDesk/AreaTime/AreaAndTimeSection.scss
@@ -1,1 +1,0 @@
-@import '../../../variables.module.scss';
