--- src/Tools/RapidResponseDesk/Provider/ProviderSection.scss
@@ -1,1 +1,0 @@
-@import '../../../variables.module.scss';
