+++ src/Tools/RapidResponseDesk/sections/Advanced/AdvancedSection.scss
@@ -1,0 +1,1 @@
+@import '../../../../variables.module';
