+++ src/Tools/RapidResponseDesk/sections/AreaTime/AreaAndTimeSection.scss
@@ -1,0 +1,1 @@
+@import '../../../../variables.module';
