+++ src/Tools/RapidResponseDesk/sections/Provider/ProviderSection.scss
@@ -1,0 +1,1 @@
+@import '../../../../variables.module';
